:recycle: Refactor inline-variable to be consistant with extract-variable
